added changes for #879 & #858
