+++ src/libs/MultiversX/dataNftMarket.ts
@@ -26,6 +26,8 @@ import jsonData from "./ABIs/data_market.abi.json";
 import { getNetworkProvider } from "./api";
 import { MarketplaceRequirementsType, OfferType } from "./types";
 import { contractsForChain } from "../config";
+import { useGetAccountInfo, useGetLoginInfo } from "@multiversx/sdk-dapp/hooks";
+import { backendApi } from "libs/utils";
 
 export class DataNftMarketContract {
   timeout: number;
@@ -246,7 +248,7 @@ export class DataNftMarketContract {
     });
   }
 
-  async delistDataNft(index: number, delistAmount: number, senderAddress: string) {
+  async delistDataNft(index: number, delistAmount: number, senderAddress: string, nativeAuthToken: string) {
     const data = new ContractCallPayloadBuilder()
       .setFunction(new ContractFunction("cancelOffer"))
       .addArg(new U64Value(index))
@@ -274,6 +276,25 @@ export class DataNftMarketContract {
       },
       redirectAfterSign: false,
     });
+
+    try {
+      const headers = {
+        Authorization: `Bearer ${nativeAuthToken}`,
+        "Content-Type": "application/json",
+      };
+
+      const requestBody = { supply: delistAmount };
+      const response = await fetch(`${backendApi(this.chainID)}/updateOffer/${index}`, {
+        method: "POST",
+        headers: headers,
+        body: JSON.stringify(requestBody),
+      });
+
+      const data = await response.json();
+      console.log("Response:", data);
+    } catch (error) {
+      console.log("Error:", error);
+    }
 
     return { sessionId, error };
   }
